access_management/requirements.txt
flask
kafka-python
SQLAlchemy
psycopg2-binary